+++ src/app/features/trackers/price-tracker/price-tracker-storage.service.ts
@@ -1,6 +1,7 @@
 import { Injectable, inject } from '@angular/core';
 import { StorageKey } from 'src/app/common/services/storage/storage';
 import { StorageService } from 'src/app/common/services/storage/storage.service';
+import { config } from 'src/config/config';
 
 export interface RecentItem {
   id: number;
@@ -13,8 +14,6 @@ export interface RecentItem {
 })
 export class PriceTrackerStorageService {
   private readonly storageService = inject(StorageService);
-
-  readonly MAX_PLAYERS_STORED = 5;
 
   getRecentItems(): RecentItem[] {
     return JSON.parse(this.storageService.getItem(StorageKey.PriceTrackerRecentItems) ?? '[]');
@@ -31,7 +30,7 @@ export class PriceTrackerStorageService {
 
     recenItems.unshift(recentItem);
 
-    if (recenItems.length > this.MAX_PLAYERS_STORED) {
+    if (recenItems.length > config.maxStoredItems) {
       recenItems.pop();
     }